Update mix.lock after running `mix local.hex`
Checking in an updated mix.lock after updating Hex locally
to the latest version. Run `mix local.hex` to make sure we're
all on the latest version and won't continue to see our local
mix.lock get out of sync with the version that's checked in.

See https://elixirforum.com/t/mix-lock-being-rewritten-with-and-then-without-hexpm/4703/2