Display handling events in correct order
Sort the handling events that are displayed in completion-time order.

Also, modify javascript to insert a new handling event received from the
web channel in the correct position in the displayed table.

Modify the show template so that it constructs unique ids for each date
and time element in a row in the table. While not strictly necessary,
this is considered good practice so that no element in a document has
the same id.